requirements.txt
pandas
gdown
mlflow==2.2.2
notebook
matplotlib
seaborn
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-PyYAML
scipy
Flask
Flask-Cors
torch
torchvision
torchaudio
-f https://download.pytorch.org/whl/cu118
-e .
